requirements.txt
streamlit==1.32.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
websocket-client==1.6.4 